backend/requirements.txt
Django>=4.0,<5.0
django-simpleui
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=4.0.0
django-filter>=23.0
python-decouple>=3.6
django-ckeditor>=6.7.0
Pillow>=10.0.0
django-redis
cryptography